+++ src/alfasim_score/converter/alfacase/_tests/test_convert_formation/test_convert_formation.yml
@@ -1,4 +1,4 @@
-- material: Folhelho
+- material: FC_Folhelho
   name: formation_1
   start:
     unit: m
